AI-chatbot/requirements.txt
# Core dependencies for StudyMate AI Assistant
streamlit>=1.28.0
PyMuPDF>=1.23.0
numpy>=1.24.0
faiss-cpu>=1.7.4
requests>=2.31.0
sentence-transformers==2.2.2
transformers==4.35.2
huggingface-hub==0.16.4
torch>=2.0.0

# Additional dependencies for enhanced functionality
pandas>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0

reportlab>=4.0.0

Pillow>=10.0.0

# Optional: OpenAI integration (requires API key)
openai>=1.0.0

# Optional: Enhanced HTTP requests for LLM APIs
urllib3>=1.26.0
